# [Java.Interop] remove Type.GetTypeInfo() calls