@jonathanpeppers 2020-01-10
Test failures are due to some behavior difference between:
```csharp
GetTypeInfo().DeclaredConstructors
```
And the new code. Figuring out the correct `BindingFlags` here. It is weird that an app still worked?